requirements.txt
asgiref==3.8.1
Django==5.0.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
pillow==10.3.0
PyJWT==2.8.0
pytz==2024.1
shortuuid==1.0.13
sqlparse==0.4.4
tzdata==2024.1
